+++ csrc/ex_guts.h
@@ -359,6 +359,9 @@ enum cforth_primitive_ids
     ID_SH_GET_INNER,
     ID_SOURCEFILENAME,
     ID_DIRNAME,
+    ID_TIME_N_DATE,
+    ID_UTIME,
+    ID_NTIME,
 #endif
 /* Add new IDs by replacing reserved IDs or extending FP routines. */
 /* Do NOT change the order of these IDs or dictionary files will break! */