+++ fth/ex_local_share/random.fs
@@ -1,0 +1,9 @@
+\ Gforth compatible random number implementation
+\ 0BSD
+
+variable seed
+
+: seed! ( n -- ) seed ! ;
+: seed-init ( -- ) utime seed! ;
+: rnd ( -- n ) 37 ;
+: random ( n -- 0..n-1 ) rnd over mod ;